requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=4.2.0=pyhd8ed1ab_0
aom=3.8.1=h078ce10_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h2aa6e3c_4
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
beautifulsoup4=4.12.3=pyha770c72_0
bleach=6.1.0=pyhd8ed1ab_0
blosc=1.21.5=hc338f07_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brunsli=0.1=h9f76cd9_0
bzip2=1.0.8=h93a5062_5
c-ares=1.26.0=h93a5062_0
c-blosc2=2.13.1=ha57e6be_0
ca-certificates=2024.2.2=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.2=pyhd8ed1ab_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py310hdcd7c05_0
charls=2.4.2=h13dd4ca_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=py310hd137fd4_0
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hb547adb_0
debugpy=1.8.0=py310h1253130_1
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
deprecated=1.2.14=pyh1a96a4e_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fonttools=4.47.2=py310hd125d64_0
fqdn=1.5.1=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
giflib=5.2.1=h1a8c8d9_3
h5py=3.10.0=nompi_py310h49e257e_101
hdf5=1.14.3=nompi_h5bb55e9_100
idna=3.6=pyhd8ed1ab_0
imagecodecs=2024.1.1=py310h5cce2fa_0
imageio=2.33.1=pyh8c1a49c_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
importlib_resources=6.1.1=pyhd8ed1ab_0
iniconfig=2.0.0=pyhd8ed1ab_0
ipykernel=6.29.0=pyh3cd1d5f_0
ipympl=0.9.3=pyhd8ed1ab_0
ipython=8.20.0=pyh707e725_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.1.1=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
jsonpointer=2.4=py310hbe9552e_3
jsonschema=4.21.1=pyhd8ed1ab_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.21.1=pyhd8ed1ab_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=5.7.1=py310hbe9552e_0
jupyter_events=0.9.0=pyhd8ed1ab_0
jupyter_server=2.12.5=pyhd8ed1ab_0
jupyter_server_terminals=0.5.2=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_0
jupyterlab_widgets=3.0.9=pyhd8ed1ab_0
jxrlib=1.1=h93a5062_3
kiwisolver=1.4.5=py310h38f39d4_1
krb5=1.21.2=h92f50d5_0
lazy_loader=0.3=pyhd8ed1ab_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libaec=1.1.2=h13dd4ca_1
libavif16=1.0.3=h76840f6_2
libblas=3.9.0=21_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=21_osxarm64_openblas
libcurl=8.5.0=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_2
libgfortran5=13.2.0=hf226fd6_2
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=21_osxarm64_openblas
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.26=openmp_h6c19121_0
libpng=1.6.39=h76d750c_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.44.2=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libtiff=4.6.0=ha8a6c65_2
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
libzopfli=1.0.3=h9f76cd9_0
llvm-openmp=17.0.6=hcd81f8e_0
lumicks.pylake=1.3.1=pyhd8ed1ab_0
lz4-c=1.9.4=hb7217d7_0
markupsafe=2.1.4=py310hd125d64_0
matplotlib-base=3.8.2=py310h9d2df84_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=1.0.0=pyhb4ecaf3_1
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=7.14.2=pyhd8ed1ab_0
nbconvert-core=7.14.2=pyhd8ed1ab_0
nbconvert-pandoc=7.14.2=pyhd8ed1ab_0
nbformat=5.9.2=pyhd8ed1ab_0
ncurses=6.4=h463b476_2
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pyhd8ed1ab_0
notebook=6.5.6=pyha770c72_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numpy=1.26.3=py310hd45542a_0
openjpeg=2.5.0=h4c1507b_3
openssl=3.2.1=h0d3ecfb_0
overrides=7.7.0=pyhd8ed1ab_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.1=py310h401b61c_0
pandoc=3.1.3=hce30654_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py310hfae7ebd_0
pip=23.3.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.0=pyhd8ed1ab_0
pluggy=1.4.0=pyhd8ed1ab_0
prometheus_client=0.19.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.8=py310hd125d64_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyobjc-core=10.1=py310hb3aa912_0
pyobjc-framework-cocoa=10.1=py310hb3aa912_0
pyparsing=3.1.1=pyhd8ed1ab_0
pytest=8.0.0=pyhd8ed1ab_0
python=3.10.8=h3ba56d0_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.19.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytz=2024.1=pyhd8ed1ab_0
pywavelets=1.4.1=py310h280b8fa_1
pyyaml=6.0.1=py310h2aa6e3c_1
pyzmq=24.0.1=py310hc407298_1
rav1e=0.6.6=h69fbcac_2
readline=8.2=h92ec313_1
referencing=0.33.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.17.1=py310hd442715_0
scikit-image=0.22.0=py310h6e3cc31_2
scikit-learn=1.4.0=py310h7ef31dd_0
scipy=1.12.0=py310hf4b343e_2
seaborn=0.12.2=py310hca03da5_0
send2trash=1.8.2=pyhd1c38e8_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=1.8.0=h463b476_0
tabulate=0.8.10=pyhd8ed1ab_0
terminado=0.18.0=pyh31c8845_0
threadpoolctl=3.2.0=pyha21a80b_0
tifffile=2024.1.30=pyhd8ed1ab_0
tikzplotlib=0.10.1=pyh050c7b8_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.3.3=py310h2aa6e3c_1
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.14.1=pyhd8ed1ab_0
types-python-dateutil=2.8.19.20240106=pyhd8ed1ab_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023d=h0c530f3_0
unicodedata2=15.1.0=py310h2aa6e3c_0
uri-template=1.3.0=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.7.0=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
widgetsnbextension=4.0.9=pyhd8ed1ab_0
wrapt=1.16.0=py310hd125d64_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=h965bd2d_0
zfp=1.0.1=ha8f4885_0
zipp=3.17.0=pyhd8ed1ab_0
zlib-ng=2.0.7=h1a8c8d9_0
zstd=1.5.5=h4f39d0f_0
